moves yaml_collector into libs/templating
