+++ openvair/libs/templating/yaml_collector.py
